requirements.txt
hydra-core==1.3.2
dotenv==0.9.9
langchain_community==0.3.27
pandas==2.3.2
langchain_upstage==0.7.3
faiss-cpu==1.12.0
qdrant-client
streamlit
python-dotenv
langsmith
langchain
langchain-openai
langchain_qdrant